Split ExternalKeyReader
